Update DataLogController.java
Not making _localDateTime null anymore since this is causing null pointer exception.

## Problem:
Since the main UI has a thread now so when this method is called it makes the _localDateTime null and then gets the date. Problem with that is if a previous all is being implemented then the _localDateTime will become null and throw a null pointer.
***
## Solution:
Make the DataLogController use a storage system to call String GetDateTimeFormatted() then this problem will be solved.
***
## Current Fix:
Currently the _localDateTime is not being nullified so this solution will work as well if garbage helper is not concerned.