Update MessageUI
Gave a thread to keep track of the length of the tweet. Tweet text turns red when tweet is over the limit. When with is over the limit it can not be added. Added _Cancel_ button to close the menu because the thread needs to be closed and the only way to do it is through another button.