Update FunnyBotGUI.java
# Fixed the file loading in linux.
#### Problem:
The file was not loading. The only part that was loading was the credentials. After that was the tweet message load that was giving problem. The problem was that when credentials were loaded properly the Datalogger was trying to log but failed. Maybe file read and write is not allowed in home folder in linux.
***
#### Solution:
Made the path of the DataLogger to load first so now the DataLogger can log in a permission given folder. This also makes sense because if any log required during loading then it will be logged in the right place.